refactor/review members segment
